Add footer label and adjust sizing and constraints of UI.
