+++ FourSix Coffee Timer/Controllers/Walkthrough/WalkthroughContentVC.swift
@@ -12,26 +12,17 @@ class WalkthroughContentVC: UIViewController, Storyboarded {
 
     @IBOutlet weak var walkthroughImage: UIImageView!
     @IBOutlet weak var headerLabel: UILabel!
+    @IBOutlet weak var footerLabel: UILabel!
 
     var headerString: String = ""
+    var footerString: String = ""
     var walkthroughImageName = ""
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
         headerLabel.text = headerString
+        footerLabel.text = footerString
         walkthroughImage.image = UIImage(named: walkthroughImageName)
     }
-
-    private func showLastPage(_ show: Bool) {
-        walkthroughImage.isHidden = show
-//        neededToolsStack.isHidden = !show
-//        endLabel.isHidden = !show
-//        startButton.isHidden = !show
-    }
-
-//    @IBAction func startTapped(_ sender: Any) {
-//        UserDefaultsManager.userHasSeenWalkthrough = true
-//        self.dismiss(animated: true)
-//    }
 }
