+++ FourSix Coffee Timer/Controllers/Walkthrough/WalkthroughModel.swift
@@ -1,0 +1,38 @@
+//
+//  WalkthroughModel.swift
+//  FourSix Coffee Timer
+//
+//  Created by Jon Duenas on 5/21/21.
+//  Copyright © 2021 Jon Duenas. All rights reserved.
+//
+
+import Foundation
+
+struct WalkthroughModel {
+    let headerStrings: [String] = [
+        "Welcome to FourSix",
+        "Creating a custom recipe tailored to your preference is as easy as 1... 2... 3...",
+        "You can preview the generated recipe as a graph and see how each pour is broken down.",
+        "Following the recipe is easy using the timer.",
+        "The new Quick Summary page gives a quick glance at how this session went.",
+        "The new Notes feature lets you add a rating and save details about your brew."
+    ]
+
+    let footerStrings: [String] = [
+        "FourSix makes brewing pour over coffee easy.",
+        "Start by setting the amount of coffee. Then choose your flavor profile and strength.",
+        "Learn how your choices change your recipe. Tap each section to see details.",
+        "Know exactly when to start pouring and how much water to pour.",
+        "Tap the notification button to get a 5 minute reminder when your coffee is at the perfect temperature.",
+        "From grind size, to water temperature, even save info about which coffee you used and who roasted it."
+    ]
+
+    let imageNames: [String] = [
+        "walkthrough-1",
+        "walkthrough-2",
+        "walkthrough-3",
+        "walkthrough-4",
+        "walkthrough-5",
+        "walkthrough-6"
+    ]
+}
