--- FourSix Coffee Timer/Model/WalkthroughModel.swift
@@ -1,29 +1,0 @@
-//
-//  WalkthroughModel.swift
-//  FourSix Coffee Timer
-//
-//  Created by Jon Duenas on 5/21/21.
-//  Copyright © 2021 Jon Duenas. All rights reserved.
-//
-
-import Foundation
-
-struct WalkthroughModel {
-    let headerStrings: [String] = [
-        "Welcome to FourSix",
-        "Start by setting the amount of coffee. Then choose your flavor profile and strength.",
-        "You can preview the generated recipe as a graph and see how each pour is broken down.",
-        "Following the recipe is easy using the timer. Know exactly when to start pouring and how much water to pour.",
-        "See the new Quick Summary page give a quick glance at how this session went.",
-        "The new Notes feature lets you add a rating and save details about your brew."
-    ]
-
-    let imageNames: [String] = [
-        "walkthrough-1",
-        "walkthrough-2",
-        "walkthrough-3",
-        "walkthrough-4",
-        "walkthrough-5",
-        "walkthrough-6"
-    ]
-}
